requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-core
langchain-google-genai
python-dotenv